+++ packages/playground/src/pages/login/login.vue
@@ -104,7 +104,7 @@ function goBack() {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "登录"
 }