+++ packages/playground/src/pages/protected/protected.vue
@@ -82,7 +82,7 @@ function goBack() {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "受保护页面",
 	"name": "ProtectedPage",