+++ packages/playground/src/pages/reports/reports.vue
@@ -377,7 +377,7 @@ onMounted(() => {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "构建报告"
 }