+++ packages/playground/src/pages/resolve/resolve.vue
@@ -120,7 +120,7 @@ function navigateTo(path: string) {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "路由解析"
 }