+++ packages/uni-app/pages-sub/profile/index.vue
@@ -128,7 +128,7 @@ export default {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "个人中心"
 }