+++ packages/uni-app/pages/error/index.vue
@@ -83,7 +83,7 @@ export default {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "错误页面"
 }