+++ packages/uni-app/pages/guards/index.vue
@@ -156,7 +156,7 @@ export default {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "路由守卫"
 }