+++ packages/uni-app/pages/navigation/index.vue
@@ -122,7 +122,7 @@ export default {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "路由导航"
 }