+++ packages/uni-app/pages/resolve/index.vue
@@ -124,7 +124,7 @@ export default {
 }
 </style>
 
-<route-config>
+<route-config lang="jsonc">
 {
 	"title": "路由解析"
 }